1.add contact page 2.change blog api address

+++ gateway/src/main/java/com/rcircle/service/gateway/services/OAuth2SsoService.java
@@ -28,30 +28,41 @@ public class OAuth2SsoService {
     @Autowired
     private RemoteSsoClient remoteSsoClient;
 
-    @HystrixCommand(fallbackMethod = "buildFallbackAccessToken", threadPoolKey = "AccessTokenThreadPool")
-    public String getAccessToken(String username, String password) {
-        HttpContextHolder.getContext().setValue(RemoteSsoRequestInterceptor.USERNAMEANDPASSWORD, username + ":" + password);
-        String state = Toolkit.randomString(8);
-        Map<String, String> parameters = new HashMap<>();
+    @HystrixCommand(fallbackMethod = "buildFallbackGetToken", threadPoolKey = "AccessTokenThreadPool")
+    private String getAuthorizeCode(Map parameters, String state){
         parameters.put("response_type", "code");
         parameters.put("client_id", clientid);
         parameters.put("redirect_uri", extractRedirect("/rst/redirect"));
         parameters.put("state", state);
-        String map = remoteSsoClient.getAuthorizeCode(parameters);
+        return remoteSsoClient.getAuthorizeCode(parameters);
+    }
+
+    @HystrixCommand(fallbackMethod = "buildFallbackGetToken", threadPoolKey = "AccessTokenThreadPool")
+    private String getToken(Map<String, String>parameters, String code){
+        parameters.remove("response");
+        parameters.put("client_secret", secret);
+        parameters.put("grant_type", "authorization_code");
+        parameters.put("code", code);
+        return remoteSsoClient.getAccessToken(parameters);
+    }
+
+    public String getAccessToken(String username, String password) {
+        String state = Toolkit.randomString(8);
+        HttpContextHolder.getContext().setValue(RemoteSsoRequestInterceptor.USERNAMEANDPASSWORD, username + ":" + password);
+        Map<String, String> parameters = new HashMap<>();
+        String map = getAuthorizeCode(parameters, state);
+        if(map.startsWith("failed")){
+            return map;
+        }
         HashMap<String, String> authcode = JSON.parseObject(map, HashMap.class);
         if (!state.equals(authcode.get("state"))) {
             return "failed! state mismatch";
         }
         HttpContextHolder.getContext().delete(RemoteSsoRequestInterceptor.USERNAMEANDPASSWORD);
-        parameters.remove("response");
-        parameters.put("client_secret", secret);
-        parameters.put("grant_type", "authorization_code");
-        parameters.put("code", authcode.get("code"));
-        String token = remoteSsoClient.getAccessToken(parameters);
-        return token;
+        return getToken(parameters, authcode.get("code"));
     }
 
-    public String buildFallbackAccessToken(String username, String password, Throwable throwable) {
+    public String buildFallbackGetToken(Map parameter, String obj, Throwable throwable) {
         int status = 0;
         String reason = null;
         if (throwable instanceof FeignException) {
